+++ src/Suru.Compiler/Types/SuruType.cs
@@ -5,4 +5,5 @@ public enum SuruType
     Bool,
     Int64,
     Float64,
+    Struct,
 }